Compensated constant time overhead
There seems to be a constant time overhead of about
1.32ms in the main loop of beep() that distorts the
frequencies (measured by recording the output and
comparing to the desired output).

This commit compensates for this overhead to output the
correct frequencies.

Note that OVERHEAD_DELAY_US was measured for this
specific implementation of beep(), so after changing
beep() might need to remeasure it.